requirements.txt
pandas>=1.5.0
google-genai>=1.0.0
opencc-python-reimplemented>=0.1.7
xlsxwriter>=3.0.0
plyer>=2.0.0
openpyxl>=3.1.0
packaging>=21.0
